Refactor/book dto

+++ microservices/library-service/src/main/java/com/livo/library_service/library/dtos/association/AssociationResponseDTO.java
@@ -2,13 +2,15 @@ package com.livo.library_service.library.dtos.association;
 
 import com.livo.library_service.library.BookStatus;
 
+import java.math.BigDecimal;
+
 public record AssociationResponseDTO(
         Long id,
         String bookId,
         BookStatus bookStatus,
         String thumbnail,
         String title,
-        Integer readingProgress,
+        BigDecimal readingProgress,
         Integer personalRatting
 ) {
 }
